Extract parallel tests into a separate file
